requirements.txt
streamlit
pandas
scikit-learn==1.3.2
xgboost
joblib
shap
matplotlib
Pillow

